Change name of Python structure solvers
