--- coupling_components/solver_wrappers/python/python_solvers.md
@@ -1,4 +1,0 @@
-# Pure Python SolverWrappers
-
-
-(empty documentation)
