+++ tests/solver_wrappers/python/test_tube_ringmodel_solver.py
@@ -1,8 +1,8 @@
 from coconut.data_structure import KratosUnittest
 
 
-class TestSolverWrapperTubeStructureSolverInert(KratosUnittest.TestCase):
-    def test_solver_wrapper_tube_structure_solver_inert(self):
+class TestSolverWrapperTubeRingmodelSolver(KratosUnittest.TestCase):
+    def test_solver_wrapper_tube_ringmodel_solver(self):
         self.assertAlmostEqual(1, 1)
